build fix "warping" flag3

+++ include/marina.h
@@ -175,7 +175,7 @@ typedef enum {
 
 s32 func_80049040(u16 actor_index);
 s32 func_80048F70(u16 actor_index);
-s32 func_800491B8(u16 actor_index, s16, s32);
+s32 func_800491B8(u16 actor_index, s32, s16);
 void ActorUpdate_Marina(u16);
 
 #endif
